refactor : Activity 관련 패키지 구조 DDD 스타일로 리팩터링
