+++ src/main/java/goormthonuniv/team_22_be/activity/domain/model/RecruitStatus.java
@@ -1,0 +1,5 @@
+package goormthonuniv.team_22_be.activity.domain.model;
+
+public enum RecruitStatus {
+    OPEN, CLOSED, FINISHED
+}
