--- src/main/java/goormthonuniv/team_22_be/api/activity/entity/RecruitStatus.java
@@ -1,5 +1,0 @@
-package goormthonuniv.team_22_be.api.activity.entity;
-
-public enum RecruitStatus {
-    OPEN, CLOSED, FINISHED
-}
