--- src/main/java/goormthonuniv/team_22_be/api/activity/repository/ActivityLikeRepository.java
@@ -1,15 +1,0 @@
-package goormthonuniv.team_22_be.api.activity.repository;
-
-import goormthonuniv.team_22_be.api.activity.entity.ActivityLike;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ActivityLikeRepository extends JpaRepository<ActivityLike, Long> {
-
-    /** 특정 회원이 특정 활동을 찜했는지 (중복 찜 방지) */
-    boolean existsByMember_IdAndActivity_Id(Long memberId, Long activityId);
-
-
-    /** 찜 해제 */
-    void deleteByMember_IdAndActivity_Id(Long memberId, Long activityId);
-
-}
